Change type of costs attribute to BigDecimal.

+++ grails-app/domain/at/uenterprise/erp/logbook/Process.groovy
@@ -10,7 +10,7 @@ class Process {
 
   String  name
   String  unit
-  Integer costs = 0
+  BigDecimal costs = 0
 
   static constraints = {
     unit  inList: ['perDay','perMonth']
